requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-dotenv==1.4.2
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
